requirements.txt
flask==2.3.3
flask-dance==6.1.0
requests==2.31.0
requests-oauthlib==1.3.1
bcrypt==4.1.2
python-dotenv==1.0.1
